backend transfer data files into the `webapi.output` package
